493 - initial commit with refactor of the ApiOperations for testing and basic tests
