+++ src/main/scala/tech/cryptonomic/conseil/routes/Data.scala
@@ -8,7 +8,7 @@ import tech.cryptonomic.conseil.config.ServerConfiguration
 import tech.cryptonomic.conseil.generic.chain.DataPlatform
 import tech.cryptonomic.conseil.generic.chain.DataTypes.QueryResponseWithOutput
 import tech.cryptonomic.conseil.metadata
-import tech.cryptonomic.conseil.metadata.{EntityPath, MetadataService, NetworkPath, Path, PlatformPath}
+import tech.cryptonomic.conseil.metadata.{EntityPath, MetadataService, NetworkPath, PlatformPath}
 import tech.cryptonomic.conseil.tezos.ApiOperations
 import tech.cryptonomic.conseil.tezos.TezosTypes.{AccountId, BlockHash}
 
@@ -16,10 +16,10 @@ import scala.concurrent.{ExecutionContext, Future}
 
 /** Companion object providing apply implementation */
 object Data {
-  def apply(config: PlatformsConfiguration, metadataService: MetadataService, server: ServerConfiguration)(
+  def apply(config: PlatformsConfiguration, metadataService: MetadataService, server: ServerConfiguration, apiOperations: ApiOperations)(
       implicit ec: ExecutionContext
   ): Data =
-    new Data(config, DataPlatform(server.maxQueryResultSize), metadataService)
+    new Data(config, DataPlatform(apiOperations, server.maxQueryResultSize), metadataService, apiOperations)
 }
 
 /**
@@ -28,7 +28,7 @@ object Data {
   * @param queryProtocolPlatform QueryProtocolPlatform object which checks if platform exists and executes query
   * @param apiExecutionContext   is used to call the async operations exposed by the api service
   */
-class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, metadataService: MetadataService)(
+class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, metadataService: MetadataService, apiOperations: ApiOperations)(
     implicit apiExecutionContext: ExecutionContext
 ) extends LazyLogging
     with DataHelpers {
@@ -70,7 +70,7 @@ class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, 
   val blocksHeadRoute: Route = blocksHeadEndpoint.implementedByAsync {
     case (platform, network, _) =>
       platformNetworkValidation(platform, network) {
-        ApiOperations.fetchLatestBlock()
+        apiOperations.fetchLatestBlock()
       }
   }
 
@@ -78,7 +78,7 @@ class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, 
   val blockByHashRoute: Route = blockByHashEndpoint.implementedByAsync {
     case ((platform, network, hash), _) =>
       platformNetworkValidation(platform, network) {
-        ApiOperations.fetchBlock(BlockHash(hash))
+        apiOperations.fetchBlock(BlockHash(hash))
       }
   }
 
@@ -94,7 +94,7 @@ class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, 
   val accountByIdRoute: Route = accountByIdEndpoint.implementedByAsync {
     case ((platform, network, accountId), _) =>
       platformNetworkValidation(platform, network) {
-        ApiOperations.fetchAccount(AccountId(accountId))
+        apiOperations.fetchAccount(AccountId(accountId))
       }
   }
 
@@ -110,7 +110,7 @@ class Data(config: PlatformsConfiguration, queryProtocolPlatform: DataPlatform, 
   val operationGroupByIdRoute: Route = operationGroupByIdEndpoint.implementedByAsync {
     case ((platform, network, operationGroupId), _) =>
       platformNetworkValidation(platform, network) {
-        ApiOperations.fetchOperationGroup(operationGroupId)
+        apiOperations.fetchOperationGroup(operationGroupId)
       }
   }
