+++ src/main/scala/tech/cryptonomic/conseil/routes/openapi/DataEndpoints.scala
@@ -1,11 +1,10 @@
 package tech.cryptonomic.conseil.routes.openapi
 
 import endpoints.algebra
-import tech.cryptonomic.conseil.generic.chain.DataTypes._
-import tech.cryptonomic.conseil.tezos.ApiOperations.{AccountResult, BlockResult, Filter, OperationGroupResult}
-import tech.cryptonomic.conseil.tezos.Tables
+import tech.cryptonomic.conseil.tezos.{Filter, Tables}
 import tech.cryptonomic.conseil.tezos.Tables.BlocksRow
-
+import tech.cryptonomic.conseil.generic.chain.DataTypes._
+import tech.cryptonomic.conseil.tezos.ApiOperations.{AccountResult, BlockResult, OperationGroupResult}
 /** Trait containing endpoints definition */
 trait DataEndpoints
     extends algebra.JsonSchemaEntities
