+++ src/test/scala/tech/cryptonomic/conseil/tezos/DataTest.scala
@@ -163,10 +163,11 @@ class DataTest
       cacheOverrides,
       platformDiscoveryOperations
     )
+  val apiOps: ApiOperations = new ApiOperations
 
-  val postRoute: Route = new Data(cfg, fakeQPP, metadataService).postRoute
+  val postRoute: Route = new Data(cfg, fakeQPP, metadataService, apiOps).postRoute
 
-  val getRoute: Route = new Data(cfg, fakeQPP, metadataService).getRoute
+  val getRoute: Route = new Data(cfg, fakeQPP, metadataService, apiOps).getRoute
 
   "Query protocol" should {
